Move bencher into test module
